requirements.txt
asgiref==3.5.0
colorama==0.4.4
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==4.0.3
django-appconf==1.0.5
django-cors-headers==3.11.0
django-cors-middleware==1.5.0
django-import-export==2.8.0
django-querycount==0.7.0
django-static-jquery==2.1.4
django-url-tools==0.0.8
django-utils-six==2.0
et-xmlfile==1.1.0
filelock==3.6.0
MarkupPy==1.14
mysqlclient @ file:///C:/Users/intern.it/Documents/Proyecto/proyectoqr/mysqlclient-2.1.0-cp310-cp310-win_amd64.whl
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.1.0
platformdirs==2.5.1
PyYAML==6.0
qrcode==7.3.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
tzdata==2022.1
virtualenv==20.14.0
virtualenvwrapper-win==1.2.7
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
